More refactoring bitcoind RPC, mostly tests
Splits more tests into their own traits. Adds util function
for creating two unconnected nodes. Useful for checking if
P2P RPC calls are working properly.
