+++ eclair-rpc/src/test/scala/org/bitcoins/rpc/BitcoindRpcTestUtil.scala
@@ -7,9 +7,16 @@ import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import org.bitcoins.core.config.RegTest
 import org.bitcoins.core.crypto.DoubleSha256Digest
+import org.bitcoins.core.currency.Bitcoins
+import org.bitcoins.core.number.UInt32
+import org.bitcoins.core.protocol.BitcoinAddress
+import org.bitcoins.core.protocol.script.ScriptSignature
+import org.bitcoins.core.protocol.transaction.{ Transaction, TransactionInput, TransactionOutPoint }
 import org.bitcoins.core.util.BitcoinSLogger
 import org.bitcoins.rpc.client.common.BitcoindRpcClient
+import org.bitcoins.rpc.client.common.RpcOpts.AddNodeArgument
 import org.bitcoins.rpc.config.{ BitcoindAuthCredentials, BitcoindInstance }
+import org.bitcoins.rpc.jsonmodels.{ GetBlockWithTransactionsResult, GetTransactionResult }
 
 import scala.concurrent.{ Await, ExecutionContext, Future, Promise }
 import scala.concurrent.duration.{ DurationInt, FiniteDuration }
@@ -200,12 +207,17 @@ trait BitcoindRpcTestUtil extends BitcoinSLogger {
       maxTries = maxTries)
   }
 
-  /** Returns a pair of RpcClients that are connected with 100 blocks in the chain */
-  def createNodePair(
+  /**
+   * Returns a pair of unconnected [[BitcoindRpcClient]]s with no blocks
+   */
+  def createUnconnectedNodePair(
     port1: Int = randomPort,
     rpcPort1: Int = randomPort,
     port2: Int = randomPort,
-    rpcPort2: Int = randomPort)(implicit system: ActorSystem): Future[(BitcoindRpcClient, BitcoindRpcClient)] = {
+    rpcPort2: Int = randomPort)(
+    implicit
+    system: ActorSystem): (BitcoindRpcClient, BitcoindRpcClient) = {
+
     implicit val m: ActorMaterializer = ActorMaterializer.create(system)
     implicit val ec: ExecutionContext = m.executionContext
     val client1: BitcoindRpcClient = new BitcoindRpcClient(instance(port1, rpcPort1))
@@ -226,27 +238,97 @@ trait BitcoindRpcTestUtil extends BitcoinSLogger {
       throw try2.failed.get
     }
 
-    client1.addNode(client2.getDaemon.uri, "add").flatMap { _ =>
-      val try3 = Try(awaitConnection(
-        from = client1,
-        to = client2,
-        duration = 1.seconds))
+    (client1, client2)
+  }
 
+  /**
+   * Returns a pair of [[BitcoindRpcClient]] that are connected with 100 blocks in the chain
+   */
+  def createNodePair(
+    port1: Int = randomPort,
+    rpcPort1: Int = randomPort,
+    port2: Int = randomPort,
+    rpcPort2: Int = randomPort)(implicit system: ActorSystem): Future[(BitcoindRpcClient, BitcoindRpcClient)] = {
+    implicit val executionContext: ExecutionContext = system.dispatcher
+    val (client1, client2) = createUnconnectedNodePair()
+    client1.addNode(client2.getDaemon.uri, AddNodeArgument.Add).map(_ => {
+      val try3 = Try(awaitConnection(client1, client2))
       if (try3.isFailure) {
         logger.error(s"Failed to connect client1 and client 2 ${try3.failed.get.getMessage}")
         deleteNodePair(client1, client2)
         throw try3.failed.get
       }
-      client1.generate(100).map { _ =>
-
-        val try4 = Try(awaitSynced(client1, client2))
-
-        if (try4.isFailure) {
-          deleteNodePair(client1, client2)
-          throw try4.failed.get
+    }).map { _ => client1.generate(100) }
+      .map { _ =>
+        {
+          val try4 = Try(awaitSynced(client1, client2))
+          if (try4.isFailure) {
+            deleteNodePair(client1, client2)
+            throw try4.failed.get
+          }
+          (client1, client2)
         }
-        (client1, client2)
       }
+  }
+
+  def createRawCoinbaseTransaction(
+    sender: BitcoindRpcClient,
+    receiver: BitcoindRpcClient,
+    amount: Bitcoins = Bitcoins(1))(implicit executionContext: ExecutionContext): Future[Transaction] = {
+    sender.generate(2).flatMap { blocks =>
+      sender.getBlock(blocks(0)).flatMap { block0 =>
+        sender.getBlock(blocks(1)).flatMap { block1 =>
+          sender.getTransaction(block0.tx(0)).flatMap { transaction0 =>
+            sender.getTransaction(block1.tx(0)).flatMap { transaction1 =>
+              val input0 =
+                TransactionOutPoint(
+                  transaction0.txid.flip,
+                  UInt32(transaction0.blockindex.get))
+              val input1 =
+                TransactionOutPoint(
+                  transaction1.txid.flip,
+                  UInt32(transaction1.blockindex.get))
+              val sig: ScriptSignature = ScriptSignature.empty
+              receiver.getNewAddress.flatMap { address =>
+                sender.createRawTransaction(
+                  Vector(
+                    TransactionInput(input0, sig, UInt32(1)),
+                    TransactionInput(input1, sig, UInt32(2))),
+                  Map(address -> amount))
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+
+  def sendCoinbaseTransaction(
+    sender: BitcoindRpcClient,
+    receiver: BitcoindRpcClient,
+    amount: Bitcoins = Bitcoins(1))(implicit executionContext: ExecutionContext): Future[GetTransactionResult] = {
+    TestUtil.createRawCoinbaseTransaction(sender, receiver, amount).flatMap {
+      transaction =>
+        sender.signRawTransaction(transaction).flatMap { signedTransaction =>
+          sender
+            .generate(100)
+            .flatMap { _ => // Can't spend coinbase until depth 100
+              sender.sendRawTransaction(
+                signedTransaction.hex,
+                allowHighFees = true).flatMap {
+                  transactionHash =>
+                    sender.getTransaction(transactionHash)
+                }
+            }
+        }
+    }
+  }
+
+  def getFirstBlock(implicit
+    node: BitcoindRpcClient,
+    executionContext: ExecutionContext): Future[GetBlockWithTransactionsResult] = {
+    node.getBlockHash(1).flatMap { hash =>
+      node.getBlockWithTransactions(hash)
     }
   }
