+++ rpc/src/test/scala/org/bitcoins/rpc/common/BitcoindRpcClientTest.scala
@@ -1,7 +1,6 @@
 package org.bitcoins.rpc.common
 
 import java.io.File
-import java.net.URI
 import java.util.Scanner
 
 import akka.actor.ActorSystem
@@ -17,14 +16,14 @@ import org.bitcoins.core.util.BitcoinSLogger
 import org.bitcoins.core.wallet.fee.SatoshisPerByte
 import org.bitcoins.rpc.client.common.RpcOpts.{ AddNodeArgument, AddressType }
 import org.bitcoins.rpc.client.common.{ BitcoindRpcClient, RpcOpts }
-import org.bitcoins.rpc.jsonmodels.{ GetBlockWithTransactionsResult, GetTransactionResult, RpcAddress }
+import org.bitcoins.rpc.jsonmodels.RpcAddress
+import org.bitcoins.rpc.{ RpcUtil, TestUtil }
 import org.scalatest.{ AsyncFlatSpec, BeforeAndAfter, BeforeAndAfterAll }
 import org.slf4j.Logger
 
 import scala.concurrent.duration.DurationInt
-import scala.concurrent.{ Await, ExecutionContext, Future }
+import scala.concurrent.{ Await, ExecutionContext }
 import scala.util.Try
-import scala.async.Async.{ async, await }
 
 class BitcoindRpcClientTest
   extends AsyncFlatSpec
@@ -35,7 +34,7 @@ class BitcoindRpcClientTest
   implicit val ec = m.executionContext
   implicit val networkParam = BitcoindRpcTestUtil.network
 
-  val client = new BitcoindRpcClient(BitcoindRpcTestUtil.instance())
+  implicit val client: BitcoindRpcClient = new BitcoindRpcClient(BitcoindRpcTestUtil.instance())
 
   val otherClient = new BitcoindRpcClient(BitcoindRpcTestUtil.instance())
 
@@ -95,59 +94,6 @@ class BitcoindRpcClientTest
     Await.result(system.terminate(), 10.seconds)
   }
 
-  private def sendCoinbaseTransaction(
-    sender: BitcoindRpcClient = client,
-    receiver: BitcoindRpcClient = otherClient,
-    amount: Bitcoins = Bitcoins(1)): Future[GetTransactionResult] = {
-    createRawCoinbaseTransaction(sender, receiver, amount).flatMap {
-      transaction =>
-        sender.signRawTransaction(transaction).flatMap { signedTransaction =>
-          sender
-            .generate(100)
-            .flatMap { _ => // Can't spend coinbase until depth 100
-              sender.sendRawTransaction(
-                signedTransaction.hex,
-                allowHighFees = true).flatMap {
-                  transactionHash =>
-                    sender.getTransaction(transactionHash)
-                }
-            }
-        }
-    }
-  }
-
-  private def createRawCoinbaseTransaction(
-    sender: BitcoindRpcClient = client,
-    receiver: BitcoindRpcClient = otherClient,
-    amount: Bitcoins = Bitcoins(1)): Future[Transaction] = {
-    sender.generate(2).flatMap { blocks =>
-      sender.getBlock(blocks(0)).flatMap { block0 =>
-        sender.getBlock(blocks(1)).flatMap { block1 =>
-          sender.getTransaction(block0.tx(0)).flatMap { transaction0 =>
-            sender.getTransaction(block1.tx(0)).flatMap { transaction1 =>
-              val input0 =
-                TransactionOutPoint(
-                  transaction0.txid.flip,
-                  UInt32(transaction0.blockindex.get))
-              val input1 =
-                TransactionOutPoint(
-                  transaction1.txid.flip,
-                  UInt32(transaction1.blockindex.get))
-              val sig: ScriptSignature = ScriptSignature.empty
-              receiver.getNewAddress.flatMap { address =>
-                sender.createRawTransaction(
-                  Vector(
-                    TransactionInput(input0, sig, UInt32(1)),
-                    TransactionInput(input1, sig, UInt32(2))),
-                  Map(address -> amount))
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
   behavior of "BitcoindRpcClient"
 
   it should "be able to prune the blockchain" in {
@@ -159,7 +105,7 @@ class BitcoindRpcClientTest
   }
 
   it should "be able to get the first block" in {
-    getFirstBlock().flatMap { block =>
+    TestUtil.getFirstBlock.flatMap { block =>
       assert(block.tx.nonEmpty)
       assert(block.height == 1)
     }
@@ -207,80 +153,22 @@ class BitcoindRpcClientTest
   it should "be able to invalidate a block" in {
     otherClient.getNewAddress.flatMap { address =>
       TestUtil.fundMemPoolTransaction(client, address, Bitcoins(3)).flatMap { txid =>
-        client.generate(1).flatMap { blocks =>
-          client.invalidateBlock(blocks.head).flatMap { _ =>
-            client.getRawMemPool.flatMap { mempool =>
-              assert(mempool.contains(txid))
-              client.getBlockCount.flatMap { count1 =>
-                otherClient.getBlockCount.flatMap { count2 =>
-                  client.generate(2) // Ensure client and otherClient have the same blockchain
-                  assert(count1 == count2 - 1)
+        client.getRawMemPool.flatMap { _ =>
+          client.generate(1).flatMap { blocks =>
+            client.invalidateBlock(blocks.head).flatMap { _ =>
+              client.getRawMemPool.flatMap { mempool =>
+                assert(mempool.contains(txid))
+                client.getBlockCount.flatMap { count1 =>
+                  otherClient.getBlockCount.flatMap { count2 =>
+                    client.generate(2) // Ensure client and otherClient have the same blockchain
+                    assert(count1 == count2 - 1)
+                  }
                 }
               }
             }
           }
         }
       }
-    }
-  }
-
-  it should "be able to get peer info" in {
-    client.getPeerInfo.flatMap { infoList =>
-      assert(infoList.length == 1)
-      val info = infoList.head
-      assert(!info.inbound)
-      assert(info.addnode)
-      assert(info.networkInfo.addr == otherClient.getDaemon.uri)
-    }
-  }
-
-  it should "be able to get the added node info" in {
-    client.getAddedNodeInfo.flatMap { info =>
-      assert(info.length == 1)
-      assert(info.head.addednode == otherClient.getDaemon.uri)
-      assert(info.head.connected.contains(true))
-    }
-  }
-
-  it should "be able to ban and clear the ban of a subnet" in {
-    val loopBack = URI.create("http://127.0.0.1")
-    BitcoindRpcTestUtil.createNodePair().flatMap {
-      case (client1, client2) =>
-        client1.setBan(loopBack, "add").flatMap { _ =>
-          client1.listBanned.flatMap { list =>
-            assert(list.length == 1)
-            assert(list.head.address.getAuthority == loopBack.getAuthority)
-            assert(
-              list.head.banned_until - list.head.ban_created == UInt32(86400))
-
-            client1.setBan(loopBack, "remove").flatMap { _ =>
-              client1.listBanned.flatMap { newList =>
-                BitcoindRpcTestUtil.deleteNodePair(client1, client2)
-                assert(newList.isEmpty)
-              }
-            }
-          }
-        }
-    }
-  }
-
-  it should "be able to clear banned subnets" in {
-    BitcoindRpcTestUtil.createNodePair().flatMap {
-      case (client1, client2) =>
-        client1.setBan(URI.create("http://127.0.0.1"), "add").flatMap { _ =>
-          client1.setBan(URI.create("http://127.0.0.2"), "add").flatMap { _ =>
-            client1.listBanned.flatMap { list =>
-              assert(list.length == 2)
-
-              client1.clearBanned().flatMap { _ =>
-                client1.listBanned.flatMap { newList =>
-                  BitcoindRpcTestUtil.deleteNodePair(client1, client2)
-                  assert(newList.isEmpty)
-                }
-              }
-            }
-          }
-        }
     }
   }
 
@@ -366,7 +254,7 @@ class BitcoindRpcClientTest
             assert(rpcAddress.address == address)
             assert(rpcAddress.balance == Bitcoins(1.25))
 
-            getFirstBlock().flatMap { block =>
+            TestUtil.getFirstBlock.flatMap { block =>
               val firstAddress =
                 block.tx.head.vout.head.scriptPubKey.addresses.get.head
               assert(
@@ -378,123 +266,6 @@ class BitcoindRpcClientTest
             }
           }
       }
-    }
-  }
-
-  it should "be able to list utxos" in {
-    client.listUnspent.flatMap { unspent =>
-      getFirstBlock().flatMap { block =>
-        val address = block.tx.head.vout.head.scriptPubKey.addresses.get.head
-        val unspentMined =
-          unspent.filter(addr => addr.address.contains(address))
-        assert(unspentMined.length >= 100)
-      }
-    }
-  }
-
-  it should "be able to lock and unlock utxos as well as list locked utxos" in {
-    client.listUnspent.flatMap { unspent =>
-      val txid1 = unspent(0).txid
-      val vout1 = unspent(0).vout
-      val txid2 = unspent(1).txid
-      val vout2 = unspent(1).vout
-      val param = Vector(
-        RpcOpts.LockUnspentOutputParameter(txid1, vout1),
-        RpcOpts.LockUnspentOutputParameter(txid2, vout2))
-      client.lockUnspent(unlock = false, param).flatMap { success =>
-        assert(success)
-        client.listLockUnspent.flatMap { locked =>
-          assert(locked.length == 2)
-          assert(locked(0).txId.flip == txid1)
-          assert(locked(1).txId.flip == txid2)
-          client.lockUnspent(unlock = true, param).flatMap { success =>
-            assert(success)
-            client.listLockUnspent.map { newLocked =>
-              assert(newLocked.isEmpty)
-            }
-          }
-        }
-      }
-    }
-  }
-
-  it should "be able to get blockchain info" in {
-    client.getBlockChainInfo.flatMap { info =>
-      assert(info.chain == "regtest")
-      assert(info.softforks.length >= 3)
-      assert(info.bip9_softforks.keySet.size >= 2)
-      client.getBestBlockHash.map(bestHash =>
-        assert(info.bestblockhash == bestHash))
-    }
-  }
-
-  it should "be able to create a raw transaction" in {
-    client.generate(2).flatMap { blocks =>
-      client.getBlock(blocks(0)).flatMap { block0 =>
-        client.getBlock(blocks(1)).flatMap { block1 =>
-          client.getTransaction(block0.tx(0)).flatMap { transaction0 =>
-            client.getTransaction(block1.tx(0)).flatMap { transaction1 =>
-              val input0 =
-                TransactionOutPoint(
-                  transaction0.txid.flip,
-                  UInt32(transaction0.blockindex.get))
-              val input1 =
-                TransactionOutPoint(
-                  transaction1.txid.flip,
-                  UInt32(transaction1.blockindex.get))
-              val sig: ScriptSignature = ScriptSignature.empty
-              otherClient.getNewAddress.flatMap { address =>
-                client
-                  .createRawTransaction(
-                    Vector(
-                      TransactionInput(input0, sig, UInt32(1)),
-                      TransactionInput(input1, sig, UInt32(2))),
-                    Map((address, Bitcoins(1))))
-                  .map { transaction =>
-                    assert(transaction.inputs.head.sequence == UInt32(1))
-                    assert(transaction.inputs(1).sequence == UInt32(2))
-                    assert(transaction
-                      .inputs.head
-                      .previousOutput
-                      .txId == input0.txId)
-                    assert(transaction
-                      .inputs(1)
-                      .previousOutput
-                      .txId == input1.txId)
-                  }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
-  it should "be able to fund a raw transaction" in {
-    otherClient.getNewAddress.flatMap { address =>
-      client
-        .createRawTransaction(Vector.empty, Map(address -> Bitcoins(1)))
-        .flatMap { transactionWithoutFunds =>
-          client.fundRawTransaction(transactionWithoutFunds).flatMap {
-            transactionResult =>
-              val transaction = transactionResult.hex
-              assert(transaction.inputs.length == 1)
-              client
-                .getRawTransaction(
-                  transaction.inputs.head.previousOutput.txId.flip)
-                .flatMap { inputTransaction =>
-                  assert(
-                    inputTransaction
-                      .vout(transaction.inputs.head.previousOutput.vout.toInt)
-                      .value
-                      .satoshis
-                      .toBigInt ==
-                      transactionResult.fee.satoshis.toBigInt +
-                      transaction.outputs.head.value.satoshis.toBigInt +
-                      transaction.outputs(1).value.satoshis.toBigInt)
-                }
-          }
-        }
     }
   }
 
@@ -548,55 +319,6 @@ class BitcoindRpcClientTest
     }
   }
 
-  it should "be able to find mem pool ancestors and descendants" in {
-    client.generate(1)
-    client.getNewAddress.flatMap { address =>
-      TestUtil.fundMemPoolTransaction(client, address, Bitcoins(2)).flatMap { txid1 =>
-        client.getRawMemPool.flatMap { mempool =>
-          assert(mempool.head == txid1)
-          client.getNewAddress.flatMap { address =>
-            val input: TransactionInput =
-              TransactionInput(
-                TransactionOutPoint(txid1.flip, UInt32(0)),
-                ScriptSignature.empty,
-                UInt32.max - UInt32.one)
-            client
-              .createRawTransaction(Vector(input), Map(address -> Bitcoins(1)))
-              .flatMap { createdTx =>
-                client.signRawTransaction(createdTx).flatMap { signedTx =>
-                  assert(signedTx.complete)
-                  client.sendRawTransaction(signedTx.hex, allowHighFees = true).flatMap {
-                    txid2 =>
-                      client.getMemPoolDescendants(txid1).flatMap {
-                        descendants =>
-                          assert(descendants.head == txid2)
-                          client
-                            .getMemPoolDescendantsVerbose(txid1)
-                            .flatMap { descendants =>
-                              assert(descendants.head._1 == txid2)
-                              assert(descendants.head._2.ancestorcount == 2)
-                              client.getMemPoolAncestors(txid2).flatMap {
-                                ancestors =>
-                                  assert(ancestors.head == txid1)
-                                  client
-                                    .getMemPoolAncestorsVerbose(txid2)
-                                    .map { ancestors =>
-                                      assert(ancestors.head._1 == txid1)
-                                      assert(
-                                        ancestors.head._2.descendantcount == 2)
-                                    }
-                              }
-                            }
-                      }
-                  }
-                }
-              }
-          }
-        }
-      }
-    }
-  }
-
   it should "be able to list transactions by receiving addresses" in {
     otherClient.getNewAddress.flatMap { address =>
       TestUtil.fundBlockChainTransaction(client, address, Bitcoins(1.5)).flatMap {
@@ -634,132 +356,8 @@ class BitcoindRpcClientTest
     }
   }
 
-  it should "be able to get a raw transaction using both rpcs available" in {
-    getFirstBlock().flatMap { block =>
-      val txid = block.tx.head.txid
-      client.getRawTransaction(txid).flatMap { transaction1 =>
-        client.getTransaction(txid).map { transaction2 =>
-          assert(transaction1.txid == transaction2.txid)
-          assert(transaction1.confirmations == transaction2.confirmations)
-          assert(transaction1.hex == transaction2.hex)
-          assert(transaction2.blockhash.contains(transaction1.blockhash))
-          assert(transaction2.blocktime.contains(transaction1.blocktime))
-        }
-      }
-    }
-  }
-
-  it should "be able to get utxo info" in {
-    getFirstBlock().flatMap { block =>
-      client.getTxOut(block.tx.head.txid, 0).map { info1 =>
-        assert(info1.coinbase)
-      }
-    }
-  }
-
-  it should "be able to decode a raw transaction" in {
-    createRawCoinbaseTransaction().flatMap { transaction =>
-      client.decodeRawTransaction(transaction).map { rpcTransaction =>
-        assert(rpcTransaction.txid == transaction.txIdBE)
-        assert(rpcTransaction.locktime == transaction.lockTime)
-        assert(rpcTransaction.size == transaction.size)
-        assert(rpcTransaction.version == transaction.version.toInt)
-        assert(rpcTransaction.vsize == transaction.vsize)
-      }
-    }
-  }
-
-  it should "be able to sign a raw transaction with wallet keys" in {
-    createRawCoinbaseTransaction().flatMap { transaction =>
-      client.signRawTransaction(transaction).map { signedTransaction =>
-        assert(signedTransaction.complete)
-      }
-    }
-  }
-
-  it should "be able to send a raw transaction to the mem pool" in {
-    createRawCoinbaseTransaction().flatMap { transaction =>
-      client.signRawTransaction(transaction).flatMap { signedTransaction =>
-        client
-          .generate(100)
-          .flatMap { _ => // Can't spend coinbase until depth 100
-            client.sendRawTransaction(signedTransaction.hex, allowHighFees = true).map {
-              _ =>
-                succeed
-            }
-          }
-      }
-    }
-  }
-
-  it should "be able to get a raw transaction in serialized form from the mem pool" in {
-    sendCoinbaseTransaction().flatMap { tx =>
-      client.getRawTransactionRaw(tx.txid).map { transaction =>
-        assert(transaction.txIdBE == tx.txid)
-      }
-    }
-  }
-
-  it should "be able to find a transaction sent to the mem pool" in {
-    sendCoinbaseTransaction().flatMap { transaction =>
-      client.getRawMemPool.map { memPool =>
-        assert(memPool.length == 1)
-        assert(memPool.head == transaction.txid)
-      }
-    }
-  }
-
-  it should "be able to find a verbose transaction in the mem pool" in {
-    sendCoinbaseTransaction().flatMap { transaction =>
-      client.getRawMemPoolWithTransactions.flatMap { memPool =>
-        val txid = memPool.keySet.head
-        assert(txid == transaction.txid)
-        assert(memPool(txid).size > 0)
-      }
-    }
-  }
-
-  it should "be able to find a mem pool entry" in {
-    sendCoinbaseTransaction().flatMap { transaction =>
-      client.getMemPoolEntry(transaction.txid).map { _ =>
-        succeed
-      }
-    }
-  }
-
-  it should "be able to get mem pool info" in {
-    client.generate(1).flatMap { _ =>
-      client.getMemPoolInfo.flatMap { info =>
-        assert(info.size == 0)
-        sendCoinbaseTransaction().flatMap { _ =>
-          client.getMemPoolInfo.map { newInfo =>
-            assert(newInfo.size == 1)
-          }
-        }
-      }
-    }
-  }
-
-  it should "be able to prioritise a mem pool transaction" in {
-    otherClient.getNewAddress.flatMap { address =>
-      TestUtil.fundMemPoolTransaction(client, address, Bitcoins(3.2)).flatMap { txid =>
-        client.getMemPoolEntry(txid).flatMap { entry =>
-          assert(entry.fee == entry.modifiedfee)
-          client.prioritiseTransaction(txid, Bitcoins(1).satoshis).flatMap {
-            tt =>
-              assert(tt)
-              client.getMemPoolEntry(txid).map { newEntry =>
-                assert(newEntry.fee == entry.fee)
-                assert(newEntry.modifiedfee == newEntry.fee + Bitcoins(1))
-              }
-          }
-        }
-      }
-    }
-  }
-
   it should "be able to get a transaction" in {
-    getFirstBlock().flatMap { block =>
+    TestUtil.getFirstBlock.flatMap { block =>
       client.getTransaction(block.tx.head.txid).flatMap { tx =>
         assert(tx.txid == block.tx.head.txid)
         assert(tx.amount == Bitcoins(50))
@@ -770,25 +368,6 @@ class BitcoindRpcClientTest
           assert(tx.confirmations == count)
         }
       }
-    }
-  }
-
-  it should "be able to get tx out proof and verify it" in {
-    getFirstBlock().flatMap { block =>
-      client.getTxOutProof(Vector(block.tx.head.txid)).flatMap { merkle =>
-        assert(merkle.transactionCount == UInt32(1))
-        assert(merkle.hashes.length == 1)
-        assert(merkle.hashes.head.flip == block.tx.head.txid)
-        client.verifyTxOutProof(merkle).map { txids =>
-          assert(block.tx.head.txid == txids.head)
-        }
-      }
-    }
-  }
-
-  it should "be able to get the network hash per sec" in {
-    client.getNetworkHashPS().map { hps =>
-      assert(hps > 0)
     }
   }
 
@@ -838,7 +417,7 @@ class BitcoindRpcClientTest
   it should "be able to get the unconfirmed balance" in { // otherClient isn't receiving txs from client???
     client.getUnconfirmedBalance.flatMap { balance =>
       assert(balance == Bitcoins(0))
-      sendCoinbaseTransaction(client, client).flatMap { transaction =>
+      TestUtil.sendCoinbaseTransaction(client, client).flatMap { transaction =>
         client.getUnconfirmedBalance.map { newBalance =>
           assert(newBalance == transaction.amount)
         }
@@ -878,10 +457,6 @@ class BitcoindRpcClientTest
     client.getConnectionCount.map { count =>
       assert(count == 1)
     }
-
-    walletClient.getConnectionCount.map { count =>
-      assert(count == 0)
-    }
   }
 
   it should "be able to get the best block hash" in {
@@ -890,69 +465,8 @@ class BitcoindRpcClientTest
     }
   }
 
-  it should "be able to get the mining info" in {
-    client.getMiningInfo.map { info =>
-      assert(info.chain == "regtest")
-    }
-  }
-
   it should "be able to get the chain tips" in {
     client.getChainTips.map { _ =>
-      succeed
-    }
-  }
-
-  it should "be able to get the network info" in {
-    client.getNetworkInfo.map { info =>
-      assert(info.networkactive)
-      assert(info.localrelay)
-      assert(info.connections == 1)
-    }
-  }
-
-  it should "be able to generate blocks" in {
-    client.generate(3).map { blocks =>
-      assert(blocks.length == 3)
-    }
-  }
-
-  it should "be able to generate blocks to an address" in {
-    otherClient.getNewAddress.flatMap { address =>
-      client.generateToAddress(3, address).flatMap { blocks =>
-        assert(blocks.length == 3)
-        blocks.foreach { hash =>
-          client.getBlockWithTransactions(hash).map { block =>
-            assert(
-              block.tx.head.vout.head.scriptPubKey.addresses.get.head == address)
-          }
-        }
-        succeed
-      }
-    }
-  }
-
-  it should "be able to generate blocks and then get their serialized headers" in {
-    client.generate(2).flatMap { blocks =>
-      client.getBlockHeaderRaw(blocks(1)).map { header =>
-        assert(header.previousBlockHashBE == blocks(0))
-      }
-    }
-  }
-
-  it should "be able to generate blocks and then get their headers" in {
-    client.generate(2).flatMap { blocks =>
-      client.getBlockHeader(blocks(0)).map { header =>
-        assert(header.nextblockhash.contains(blocks(1)))
-      }
-      client.getBlockHeader(blocks(1)).map { header =>
-        assert(header.previousblockhash.contains(blocks(0)))
-        assert(header.nextblockhash.isEmpty)
-      }
-    }
-  }
-
-  it should "be able to get a block template" in {
-    client.getBlockTemplate().map { _ =>
       succeed
     }
   }
@@ -988,26 +502,6 @@ class BitcoindRpcClientTest
         assert(genHelp != helpHelp)
         assert(!helpHelp.isEmpty)
       }
-    }
-  }
-
-  it should "be able to deactivate and activate the network" in {
-    client.setNetworkActive(false).flatMap { _ =>
-      client.getNetworkInfo.flatMap { info =>
-        assert(!info.networkactive)
-        client.setNetworkActive(true).flatMap { _ =>
-          client.getNetworkInfo.map { newInfo =>
-            assert(newInfo.networkactive)
-          }
-        }
-      }
-    }
-  }
-
-  it should "be able to get the difficulty on the network" in {
-    client.getDifficulty.map { difficulty =>
-      assert(difficulty > 0)
-      assert(difficulty < 1)
     }
   }
 
@@ -1066,10 +560,6 @@ class BitcoindRpcClientTest
     }
   }
 
-  it should "be able to ping" in {
-    client.ping().map(_ => succeed)
-  }
-
   it should "be able to validate a bitcoin address" in {
     otherClient.getNewAddress.flatMap { address =>
       client.validateAddress(address).map { validation =>
@@ -1081,54 +571,6 @@ class BitcoindRpcClientTest
   it should "be able to verify the chain" in {
     client.verifyChain(blocks = 0).map { valid =>
       assert(valid)
-    }
-  }
-
-  it should "be able to get the memory info" in {
-    client.getMemoryInfo.map { info =>
-      assert(info.locked.used > 0)
-      assert(info.locked.free > 0)
-      assert(info.locked.total > 0)
-      assert(info.locked.locked > 0)
-      assert(info.locked.chunks_used > 0)
-    }
-  }
-
-  it should "be able to get network statistics" in {
-    client.getNetTotals.map { stats =>
-      assert(stats.timemillis.toBigInt > 0)
-      assert(stats.totalbytesrecv > 0)
-      assert(stats.totalbytessent > 0)
-    }
-  }
-
-  it should "be able to create a multi sig address" in {
-    val ecPrivKey1 = ECPrivateKey.freshPrivateKey
-    val ecPrivKey2 = ECPrivateKey.freshPrivateKey
-
-    val pubKey1 = ecPrivKey1.publicKey
-    val pubKey2 = ecPrivKey2.publicKey
-
-    client.createMultiSig(2, Vector(pubKey1, pubKey2)).map { _ =>
-      succeed
-    }
-    succeed
-  }
-
-  it should "be able to add a multi sig address to the wallet" in {
-    val ecPrivKey1 = ECPrivateKey.freshPrivateKey
-    val pubKey1 = ecPrivKey1.publicKey
-
-    client.getNewAddress(addressType = AddressType.Legacy).flatMap { address =>
-      client
-        .addMultiSigAddress(
-          2,
-          Vector(
-            Left(pubKey1),
-            Right(address.asInstanceOf[P2PKHAddress])))
-        .map { _ =>
-          succeed
-        }
     }
   }
 
@@ -1224,80 +666,6 @@ class BitcoindRpcClientTest
       }
   }
 
-  it should "be able to dump the wallet" in {
-    client
-      .dumpWallet(client.getDaemon.authCredentials.datadir + "/test.dat")
-      .map { result =>
-        assert(result.filename.exists)
-        assert(result.filename.isFile)
-      }
-  }
-
-  it should "be able to backup the wallet" in {
-    client
-      .backupWallet(client.getDaemon.authCredentials.datadir + "/backup.dat")
-      .map { _ =>
-        val file =
-          new File(client.getDaemon.authCredentials.datadir + "/backup.dat")
-        assert(file.exists)
-        assert(file.isFile)
-      }
-  }
-
-  it should "be able to lock and unlock the wallet" in {
-    walletClient.walletLock().flatMap { _ =>
-      walletClient.walletPassphrase(password, 1000).flatMap { _ =>
-        walletClient.getWalletInfo.flatMap { info =>
-          assert(info.unlocked_until.nonEmpty)
-          assert(info.unlocked_until.get > 0)
-
-          walletClient.walletLock().flatMap { _ =>
-            walletClient.getWalletInfo.map { newInfo =>
-              assert(newInfo.unlocked_until.contains(0))
-            }
-          }
-        }
-      }
-    }
-  }
-
-  it should "be able to change the wallet password - refactored" in async {
-    val newPass = "new_password"
-
-    await(walletClient.walletLock())
-    await(walletClient.walletPassphraseChange(password, newPass))
-    password = newPass
-    await(walletClient.walletPassphrase(password, 1000))
-    val info = await(walletClient.getWalletInfo)
-    assert(info.unlocked_until.nonEmpty)
-    assert(info.unlocked_until.get > 0)
-    await(walletClient.walletLock())
-    val newInfo = await(walletClient.getWalletInfo)
-    assert(newInfo.unlocked_until.contains(0))
-  }
-
-  it should "be able to change the wallet password" in {
-    val newPass = "new_password"
-    walletClient.walletLock().flatMap { _ =>
-      walletClient.walletPassphraseChange(password, newPass).flatMap {
-        _ =>
-          password = newPass
-          walletClient.walletPassphrase(password, 1000).flatMap { _ =>
-            walletClient.getWalletInfo.flatMap { info =>
-              assert(info.unlocked_until.nonEmpty)
-              assert(info.unlocked_until.get > 0)
-
-              walletClient.walletLock().flatMap { _ =>
-                walletClient.getWalletInfo.map { newInfo =>
-                  assert(newInfo.unlocked_until.contains(0))
-                }
-              }
-            }
-          }
-      }
-    }
-  }
-
   it should "be able to import a wallet" in {
     client.getNewAddress.flatMap { address =>
       client
@@ -1318,12 +686,6 @@ class BitcoindRpcClientTest
               }
           }
         }
-    }
-  }
-
-  it should "be able to list wallets" in {
-    client.listWallets.map { wallets =>
-      assert(wallets == Vector("wallet.dat"))
     }
   }
 
@@ -1561,13 +923,6 @@ class BitcoindRpcClientTest
     }
   }
 
-  it should "be able to abort a rescan of the blockchain" in {
-    recoverToSucceededIf[RuntimeException](client.rescanBlockChain())
-    client.abortRescan().map { _ =>
-      succeed
-    }
-  }
-
   it should "be able to add and remove a node" in {
     otherClient.addNode(walletClient.getDaemon.uri, AddNodeArgument.Add).flatMap { _ =>
       BitcoindRpcTestUtil.awaitConnection(otherClient, walletClient)
@@ -1599,13 +954,6 @@ class BitcoindRpcClientTest
           }
         }
       }
-    }
-  }
-
-  private def getFirstBlock(
-    node: BitcoindRpcClient = client): Future[GetBlockWithTransactionsResult] = {
-    node.getBlockHash(1).flatMap { hash =>
-      node.getBlockWithTransactions(hash)
     }
   }
 }
