+++ rpc/src/test/scala/org/bitcoins/rpc/common/BlockchainRpcTest.scala
@@ -1,0 +1,81 @@
+package org.bitcoins.rpc.common
+
+import akka.actor.ActorSystem
+import akka.stream.ActorMaterializer
+import org.bitcoins.core.config.NetworkParameters
+import org.bitcoins.core.currency.Bitcoins
+import org.bitcoins.core.number.UInt32
+import org.bitcoins.rpc.TestUtil
+import org.bitcoins.rpc.client.common.BitcoindRpcClient
+import org.bitcoins.rpc.client.common.RpcOpts.{ AddNodeArgument, AddressType }
+import org.scalatest.{ AsyncFlatSpec, BeforeAndAfterAll }
+
+import scala.concurrent.{ Await, ExecutionContext, Future }
+import scala.async.Async.{ async, await }
+import scala.concurrent.duration.DurationInt
+
+class BlockchainRpcTest extends AsyncFlatSpec with BeforeAndAfterAll {
+
+  implicit val system: ActorSystem = ActorSystem("RpcClientTest_ActorSystem")
+  implicit val m: ActorMaterializer = ActorMaterializer()
+  implicit val ec: ExecutionContext = m.executionContext
+  implicit val networkParam: NetworkParameters = TestUtil.network
+
+  implicit val client: BitcoindRpcClient = new BitcoindRpcClient(TestUtil.instance())
+  val otherClient = new BitcoindRpcClient(TestUtil.instance())
+
+  override def beforeAll(): Unit = {
+    TestUtil.startServers(client, otherClient)
+    Await.result(client.addNode(otherClient.getDaemon.uri, AddNodeArgument.Add), 3.seconds)
+    Await.result(client.generate(200), 3.seconds)
+    TestUtil.awaitConnection(client, otherClient)
+  }
+
+  override def afterAll(): Unit = {
+    TestUtil.stopServers(client, otherClient)
+    Await.result(system.terminate(), 10.seconds)
+  }
+
+  behavior of "BlockchainRpc"
+
+  it should "be able to get blockchain info" in {
+    client.getBlockChainInfo.flatMap { info =>
+      assert(info.chain == "regtest")
+      assert(info.softforks.length >= 3)
+      assert(info.bip9_softforks.keySet.size >= 2)
+      client.getBestBlockHash.map(bestHash =>
+        assert(info.bestblockhash == bestHash))
+    }
+  }
+
+  it should "be able to invalidate a block" in async {
+    await(client.getBalance)
+    val address = await(otherClient.getNewAddress(addressType = AddressType.P2SHSegwit))
+    val txid = await(client.sendToAddress(address, Bitcoins(1)))
+    val oldMempool = await(client.getRawMemPool)
+    val blocks = await(client.generate(1))
+    val mostRecentBlock = await(client.getBlock(blocks.head))
+    assert(mostRecentBlock.tx.contains(txid))
+    await(client.invalidateBlock(blocks.head))
+    val mempool = await(client.getRawMemPool)
+    assert(mempool.contains(txid))
+    val count1 = await(client.getBlockCount)
+    val count2 = await(otherClient.getBlockCount)
+    client.generate(2) // Ensure client and otherClient have the same blockchain
+    assert(count1 == count2 - 1)
+  }
+
+  it should "be able to get tx out proof and verify it" in {
+    TestUtil.getFirstBlock.flatMap { block =>
+      client.getTxOutProof(Vector(block.tx.head.txid)).flatMap { merkle =>
+        assert(merkle.transactionCount == UInt32(1))
+        assert(merkle.hashes.length == 1)
+        assert(merkle.hashes.head.flip == block.tx.head.txid)
+        client.verifyTxOutProof(merkle).map { txids =>
+          assert(block.tx.head.txid == txids.head)
+        }
+      }
+    }
+  }
+
+}
